backend/requirements.txt
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jsonpickle==3.0.4
jsonpointer==2.4
Markdown==3.7
msgpack==1.1.0
mysqlclient==2.2.4
packaging==24.2
paypal-server-sdk==0.6.0
paypalrestsdk==1.13.3
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.5.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.1
typing==3.7.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
